server/requirements.txt
alembic==1.7.7
anyio==3.5.0
appnope==0.1.3
asgiref==3.5.0
asttokens==2.0.5
async-generator==1.10
asyncpg==0.25.0
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.1
databases==0.5.5
debugpy==1.6.0
decorator==5.1.1
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.0
entrypoints==0.4
executing==0.8.3
fastapi==0.75.2
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore
httptools==0.4.0
httpx==0.23.0
idna==3.3
iniconfig==1.1.1
ipykernel==6.13.1
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-client==7.3.4
jupyter-core==4.10.0
limits==1.6
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
munch==2.5.0
nest-asyncio==1.5.5
numpy==1.22.3
orjson==3.6.8
outcome==1.1.0
packaging==21.3
pandas==1.4.2
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.29
psutil==5.9.1
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
Pygments==2.12.0
pyparsing==3.0.9
pytest==7.1.2
pytest-asyncio==0.18.3
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
pyzmq==23.1.0
redis==3.5.3
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
six==1.16.0
slowapi==0.1.5
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.36
sqlalchemy2-stubs==0.0.2a22
sqlmodel==0.0.6
stack-data==0.2.0
starlette==0.17.1
structlog==21.5.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
traitlets==5.2.2.post1
trio==0.20.0
typing-extensions==4.2.0
ujson==5.2.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
wcwidth==0.2.5
websockets==10.3
